be/app/requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
markdown==3.4.3
psycopg2-binary==2.9.6
sentry-sdk==1.21.1